requirements.txt
PyYAML==6.0.2
tabulate==0.9.0
